完成两次作业
