+++ src/controller/AccountServlet.java
@@ -1,14 +1,6 @@
 package controller;
 
-import service.AccountService;
-import service.User;
-
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import java.io.IOException;
 
 
 /**
