+++ src/controller/LoginServlet.java
@@ -1,7 +1,7 @@
 package controller;
 
 import service.AccountService;
-import service.User;
+import bean.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
