+++ web/rooter/rooter.admin.jsp
@@ -1,5 +1,5 @@
 <%@ page import="service.AccountService" %>
-<%@ page import="service.User" %><%--
+<%@ page import="bean.User" %><%--
   Created by IntelliJ IDEA.
   User: asus
   Date: 2019/7/16
